+++ presenter-api/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/notes/NotesBarService.java
@@ -1,0 +1,10 @@
+package ru.s4m1d.notes.app.desktop.client.presenter.notes;
+
+import ru.s4m1d.notes.app.desktop.client.model.note.Note;
+
+import java.util.List;
+
+public interface NotesBarService {
+    public void addNotes(List<Note> notes);
+    public void addNote(Note note);
+}
