+++ presenter-api/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/workspace/WorkspaceService.java
@@ -1,0 +1,4 @@
+package ru.s4m1d.notes.app.desktop.client.presenter.workspace;
+
+public interface WorkspaceService {
+}
