+++ presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/notes/NotesBarServiceObservable.java
@@ -2,5 +2,5 @@ package ru.s4m1d.notes.app.desktop.client.presenter.notes;
 
 import ru.s4m1d.notes.app.desktop.client.core.Observable;
 
-public class NotesMonitoringServiceObservable extends Observable {
+public class NotesBarServiceObservable extends Observable {
 }
