--- presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/notes/NotesBarUpdateService.java
@@ -1,4 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.presenter.notes;
-
-public interface NotesBarUpdateService {
-}
