--- presenter/src/main/java/ru/s4m1d/notes/app/desktop/client/presenter/notes/NotesBarUpdateServiceObservable.java
@@ -1,6 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.presenter.notes;
-
-import ru.s4m1d.notes.app.desktop.client.core.Observable;
-
-public class NotesBarUpdateServiceObservable extends Observable {
-}
